Move HudderUtils.java to utils package
