+++ src/main/java/io/github/ngspace/hudder/compilers/HudderV2Compiler.java
@@ -5,8 +5,8 @@ import java.util.Arrays;
 import io.github.ngspace.hudder.Hudder;
 import io.github.ngspace.hudder.compilers.abstractions.AV2Compiler;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
-import io.github.ngspace.hudder.main.HudderUtils;
 import io.github.ngspace.hudder.main.config.HudderConfig;
+import io.github.ngspace.hudder.utils.HudderUtils;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
 import io.github.ngspace.hudder.v2runtime.runtime_elements.BreakV2RuntimeElement;
 import io.github.ngspace.hudder.v2runtime.runtime_elements.ConditionV2RuntimeElement;
