+++ src/main/java/io/github/ngspace/hudder/v2runtime/values/DefaultV2VariableParser.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 
 import io.github.ngspace.hudder.compilers.abstractions.AV2Compiler;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
-import io.github.ngspace.hudder.main.HudderUtils;
+import io.github.ngspace.hudder.utils.HudderUtils;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
 import io.github.ngspace.hudder.v2runtime.values.constants.V2Array;
 import io.github.ngspace.hudder.v2runtime.values.constants.V2Boolean;
